Move color conversions into their own file
